requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
Pillow==10.4.0
gTTS==2.5.3
pydub==0.25.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
sentence-transformers==3.0.1
huggingface-hub==0.24.6
python-dotenv==1.0.1
imageio==2.34.2
imageio-ffmpeg==0.5.1
scipy==1.13.1
